Add Parser-Classes for Wifi-QR-String
As the app gets fed with QR-Codes that translate to strings, we have
to be able to validate and extract infos from the standardized format.
Therefore we implemented logic to tokenize the strings and check if
necessary attributes are set and appropriate. From there we can
re-generate the String, check if this is a Wifi-Configuration for a
Magic-Mirror or generate a WifiConfiguration-Object for usage with the
Android WifiManager.

- Add ability to validate and parse QR-Code-Strings
- Add checks to identify suitable Mirror-Configurations
- Add ability to generate WifiConfiguration for the WifiManager